job-assistant-agent/pyproject.toml
[project]
name = "job-assistant-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.23",
    "langchain-google-genai>=2.1.3",
    "litellm>=1.69.3",
    "openai-agents>=0.0.12",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
